patch core checkouts in addition to develop checkouts
Deprecate calculating level, we are always working from level 0.  Instead, run the command from either the root of our WP checkout or from the src directory. fixes #3

Add a bunch of example diffs along with a bunch of tests for calculating if we should run patch from the src directory see #4