Add password management for ticket uploads
Uses the Clortho NPM package to securely save username and password
for authentication when uploading patches.

If a username doesn't have a password saved, it will prompt to save
it using either native dialog boxes, or the CLI. Afterwards, the user
will be prompted for a WP.org username, and the saved password will
be automatically used without further prompting.