Add upstream Props Bot improvements.
This syncs over upstream improvements to the Props Bot example workflow.

The important change in this commit is ensuring the workflow does not run on the `issue_comment` event when a PR is not the subject of the comment. Previously, the conditions were too loose, allowing any event to run as long as `labeled` was not the action.